+++ watch/dvc/__init__.py
@@ -1,0 +1,6 @@
+"""
+THIS IS AN EXPERIMENTAL MODULE.
+
+The goal of this submodule is to provide tooling to create, manage, and
+interact with DVC respositories of datasets and experiments.
+"""