+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Can now force nodata values in prep ta2 and align script.
 * Perceiver backbone
 * New "window_space_scale" parameter to Dataset that allows a WINDOW GSD to be given on the fly.
+* New "watch.dvc" module for phase2 dvc operations.
 
 ### Changed
 * Data loader now returns the size it would like the output predicted at via `target_dims`.